Move parseReactionShorthand into editor/lib/emoji
